I: organize some comopnents
