--- pages/normal.js
@@ -1,3 +1,0 @@
-export default function NormalPage() {
-  return <p>I'm just a normal old page, no AMP for me</p>
-}
